fix: fix bonus point of user skill service

+++ src/main/java/com/hcmus/mela/skills/dto/UserSkillDto.java
@@ -9,5 +9,5 @@ import lombok.*;
 @AllArgsConstructor
 public class UserSkillDto {
     private String topicName;
-    private Double points;
+    private Integer points;
 }
